Apply l2 normalization to fasttext field vectors
Instances of MatrixSimilarity classes have an attribute 'normalize' and it's true for our field vectors. This results in normalization of the input document vector before similarity calculations. The field vectors are not normalized; those have to be to normalized on creation. I did this for the fasttext entity field vectors, but not the fasttext entity vectors.

The effect of this seems to be an underweighting of the fasttext field vectors in scoring, relative to the fasttext *entity* field vectors and tf-idf vectors.