Add field_meta, field_hierarchy, and field_children to the set of tables copied from staging to production on DAG run
The field_meta and field_children tables are largely static. They can be updated by making changes in the JSONL files in assets/fields and running upload.sh, which loads them into staging. The field_hierarchy table, then, is written from a DAG query using these metadata tables. All three need to be copied into the production dataset for it to be consistent.